Move support helpers from OpenAIConfig to SupportConfig
